requirements.txt
requests
python-dotenv
watchdog
boto3
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
